refactor: break up actorstate tasks per model extracted
- each actor extractor is responsible for producing a single model
- add tracing throughout actor tasks
