--- tasks/actorstate/account.go
@@ -1,24 +1,0 @@
-package actorstate
-
-import (
-	"context"
-
-	"github.com/filecoin-project/lily/chain/actors/builtin/account"
-	"github.com/filecoin-project/lily/model"
-)
-
-// AccountExtractor is a state extractor that deals with Account actors.
-type AccountExtractor struct{}
-
-func init() {
-	for _, c := range account.AllCodes() {
-		Register(c, AccountExtractor{})
-	}
-}
-
-// Extract will create persistable data for a given actor's state.
-func (AccountExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAPI) (model.Persistable, error) {
-	return model.NoData, nil
-}
-
-var _ ActorStateExtractor = (*AccountExtractor)(nil)
