+++ tasks/actorstate/init_/init.go
@@ -1,4 +1,4 @@
-package actorstate
+package init_
 
 import (
 	"context"
@@ -6,35 +6,36 @@ import (
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/specs-actors/actors/builtin"
+	logging "github.com/ipfs/go-log/v2"
 	"go.opentelemetry.io/otel"
+	"go.uber.org/zap"
 	"golang.org/x/xerrors"
 
 	init_ "github.com/filecoin-project/lily/chain/actors/builtin/init"
 	"github.com/filecoin-project/lily/metrics"
 	"github.com/filecoin-project/lily/model"
 	initmodel "github.com/filecoin-project/lily/model/actors/init"
+	"github.com/filecoin-project/lily/tasks/actorstate"
 )
 
-// was services/processor/tasks/init/init_actor.go
+var log = logging.Logger("lily/tasks/init")
 
 // InitExtractor extracts init actor state
 type InitExtractor struct{}
 
-func init() {
-	for _, c := range init_.AllCodes() {
-		Register(c, InitExtractor{})
-	}
-}
-
-func (InitExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAPI) (model.Persistable, error) {
-	ctx, span := otel.Tracer("").Start(ctx, "InitExtractor")
+func (InitExtractor) Extract(ctx context.Context, a actorstate.ActorInfo, node actorstate.ActorStateAPI) (model.Persistable, error) {
+	log.Debugw("extract", zap.String("extractor", "InitExtractor"), zap.Inline(a))
+	ctx, span := otel.Tracer("").Start(ctx, "InitExtractor.Extract")
 	defer span.End()
+	if span.IsRecording() {
+		span.SetAttributes(a.Attributes()...)
+	}
 
 	stop := metrics.Timer(ctx, metrics.StateExtractionDuration)
 	defer stop()
 
 	// genesis state.
-	if a.Epoch == 1 {
+	if a.Current.Height() == 1 {
 		initActorState, err := init_.Load(node.Store(), &a.Actor)
 		if err != nil {
 			return nil, err
@@ -50,7 +51,7 @@ func (InitExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAP
 				Height:    0,
 				ID:        builtinAddress.String(),
 				Address:   builtinAddress.String(),
-				StateRoot: a.ParentTipSet.ParentState().String(),
+				StateRoot: a.Executed.ParentState().String(),
 			})
 		}
 		if err := initActorState.ForEachActor(func(id abi.ActorID, addr address.Address) error {
@@ -59,10 +60,10 @@ func (InitExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAP
 				return err
 			}
 			out = append(out, &initmodel.IdAddress{
-				Height:    int64(a.Epoch),
+				Height:    int64(a.Current.Height()),
 				ID:        idAddr.String(),
 				Address:   addr.String(),
-				StateRoot: a.ParentStateRoot.String(),
+				StateRoot: a.Current.ParentState().String(),
 			})
 			return nil
 		}); err != nil {
@@ -70,7 +71,7 @@ func (InitExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAP
 		}
 		return out, nil
 	}
-	prevActor, err := node.Actor(ctx, a.Address, a.ParentTipSet.Key())
+	prevActor, err := node.Actor(ctx, a.Address, a.Executed.Key())
 	if err != nil {
 		return nil, xerrors.Errorf("loading previous init actor: %w", err)
 	}
@@ -93,16 +94,16 @@ func (InitExtractor) Extract(ctx context.Context, a ActorInfo, node ActorStateAP
 	out := make(initmodel.IdAddressList, 0, len(addressChanges.Added)+len(addressChanges.Modified))
 	for _, newAddr := range addressChanges.Added {
 		out = append(out, &initmodel.IdAddress{
-			Height:    int64(a.Epoch),
-			StateRoot: a.ParentStateRoot.String(),
+			Height:    int64(a.Current.Height()),
+			StateRoot: a.Current.ParentState().String(),
 			ID:        newAddr.ID.String(),
 			Address:   newAddr.PK.String(),
 		})
 	}
 	for _, modAddr := range addressChanges.Modified {
 		out = append(out, &initmodel.IdAddress{
-			Height:    int64(a.Epoch),
-			StateRoot: a.ParentStateRoot.String(),
+			Height:    int64(a.Current.Height()),
+			StateRoot: a.Current.ParentState().String(),
 			ID:        modAddr.To.ID.String(),
 			Address:   modAddr.To.PK.String(),
 		})
