+++ tasks/actorstate/miner/deadline_info.go
@@ -1,0 +1,56 @@
+package miner
+
+import (
+	"context"
+
+	"go.opentelemetry.io/otel"
+	"go.uber.org/zap"
+	"golang.org/x/xerrors"
+
+	"github.com/filecoin-project/lily/model"
+	minermodel "github.com/filecoin-project/lily/model/actors/miner"
+	"github.com/filecoin-project/lily/tasks/actorstate"
+)
+
+type DeadlineInfoExtractor struct{}
+
+func (DeadlineInfoExtractor) Extract(ctx context.Context, a actorstate.ActorInfo, node actorstate.ActorStateAPI) (model.Persistable, error) {
+	log.Debugw("extract", zap.String("extractor", "DeadlineInfoExtractor"), zap.Inline(a))
+	ctx, span := otel.Tracer("").Start(ctx, "DeadlineInfoExtractor.Extract")
+	defer span.End()
+	if span.IsRecording() {
+		span.SetAttributes(a.Attributes()...)
+	}
+
+	ec, err := NewMinerStateExtractionContext(ctx, a, node)
+	if err != nil {
+		return nil, xerrors.Errorf("creating miner state extraction context: %w", err)
+	}
+	currDeadlineInfo, err := ec.CurrState.DeadlineInfo(ec.CurrTs.Height())
+	if err != nil {
+		return nil, err
+	}
+
+	if ec.HasPreviousState() {
+		prevDeadlineInfo, err := ec.PrevState.DeadlineInfo(ec.CurrTs.Height())
+		if err != nil {
+			return nil, err
+		}
+		if prevDeadlineInfo == currDeadlineInfo {
+			return nil, nil
+		}
+	}
+
+	return &minermodel.MinerCurrentDeadlineInfo{
+		Height:        int64(ec.CurrTs.Height()),
+		MinerID:       a.Address.String(),
+		StateRoot:     a.Current.ParentState().String(),
+		DeadlineIndex: currDeadlineInfo.Index,
+		PeriodStart:   int64(currDeadlineInfo.PeriodStart),
+		Open:          int64(currDeadlineInfo.Open),
+		Close:         int64(currDeadlineInfo.Close),
+		Challenge:     int64(currDeadlineInfo.Challenge),
+		FaultCutoff:   int64(currDeadlineInfo.FaultCutoff),
+	}, nil
+
+}
