+++ tasks/actorstate/miner/info.go
@@ -1,0 +1,88 @@
+package miner
+
+import (
+	"context"
+
+	"github.com/filecoin-project/go-address"
+	logging "github.com/ipfs/go-log/v2"
+	maddr "github.com/multiformats/go-multiaddr"
+	"go.opentelemetry.io/otel"
+	"go.uber.org/zap"
+	"golang.org/x/xerrors"
+
+	"github.com/filecoin-project/lily/model"
+	minermodel "github.com/filecoin-project/lily/model/actors/miner"
+	"github.com/filecoin-project/lily/tasks/actorstate"
+)
+
+var log = logging.Logger("lily/tasks/miner")
+
+type InfoExtractor struct{}
+
+func (InfoExtractor) Extract(ctx context.Context, a actorstate.ActorInfo, node actorstate.ActorStateAPI) (model.Persistable, error) {
+	log.Debugw("extract", zap.String("extractor", "InfoExtractor"), zap.Inline(a))
+	ctx, span := otel.Tracer("").Start(ctx, "InfoExtractor.Extract")
+	defer span.End()
+	if span.IsRecording() {
+		span.SetAttributes(a.Attributes()...)
+	}
+	ec, err := NewMinerStateExtractionContext(ctx, a, node)
+	if err != nil {
+		return nil, xerrors.Errorf("creating miner state extraction context: %w", err)
+	}
+
+	if !ec.HasPreviousState() {
+		// means this miner was created in this tipset or genesis special case
+	} else if changed, err := ec.CurrState.MinerInfoChanged(ec.PrevState); err != nil {
+		return nil, err
+	} else if !changed {
+		return nil, nil
+	}
+	// miner info has changed.
+
+	newInfo, err := ec.CurrState.Info()
+	if err != nil {
+		return nil, err
+	}
+
+	var newWorker string
+	if newInfo.NewWorker != address.Undef {
+		newWorker = newInfo.NewWorker.String()
+	}
+
+	var newCtrlAddresses []string
+	for _, addr := range newInfo.ControlAddresses {
+		newCtrlAddresses = append(newCtrlAddresses, addr.String())
+	}
+
+	// best effort to decode, we have no control over what miners put in this field, its just bytes.
+	var newMultiAddrs []string
+	for _, addr := range newInfo.Multiaddrs {
+		newMaddr, err := maddr.NewMultiaddrBytes(addr)
+		if err == nil {
+			newMultiAddrs = append(newMultiAddrs, newMaddr.String())
+		} else {
+			// TODO(frrist)
+			log.Debugw("failed to decode miner multiaddr", "miner", a.Address, "multiaddress", addr, "error", err)
+		}
+	}
+	mi := &minermodel.MinerInfo{
+		Height:                  int64(ec.CurrTs.Height()),
+		MinerID:                 a.Address.String(),
+		StateRoot:               a.Current.ParentState().String(),
+		OwnerID:                 newInfo.Owner.String(),
+		WorkerID:                newInfo.Worker.String(),
+		NewWorker:               newWorker,
+		WorkerChangeEpoch:       int64(newInfo.WorkerChangeEpoch),
+		ConsensusFaultedElapsed: int64(newInfo.ConsensusFaultElapsed),
+		ControlAddresses:        newCtrlAddresses,
+		MultiAddresses:          newMultiAddrs,
+		SectorSize:              uint64(newInfo.SectorSize),
+	}
+
+	if newInfo.PeerId != nil {
+		mi.PeerID = newInfo.PeerId.String()
+	}
+
+	return mi, nil
+}
