+++ tasks/actorstate/task.go
@@ -3,12 +3,14 @@ package actorstate
 import (
 	"context"
 	"fmt"
+	"sync"
 	"time"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/lotus/chain/types"
 	"github.com/ipfs/go-cid"
 	"go.opencensus.io/tag"
+	"go.opentelemetry.io/otel"
 	"golang.org/x/xerrors"
 
 	"github.com/filecoin-project/lily/chain/actors/builtin"
@@ -23,6 +25,8 @@ type Task struct {
 	node tasks.DataSource
 
 	extracterMap ActorExtractorMap
+
+	apwg sync.WaitGroup
 }
 
 func NewTask(node tasks.DataSource, extracterMap ActorExtractorMap) *Task {
@@ -33,28 +37,28 @@ func NewTask(node tasks.DataSource, extracterMap ActorExtractorMap) *Task {
 	return p
 }
 
-func (t *Task) ProcessActors(ctx context.Context, ts *types.TipSet, pts *types.TipSet, candidates tasks.ActorStateChangeDiff) (model.Persistable, *visormodel.ProcessingReport, error) {
-	log.Debugw("processing actor state changes", "height", ts.Height(), "parent_height", pts.Height())
-
+func (t *Task) ProcessActors(ctx context.Context, current *types.TipSet, executed *types.TipSet, candidates tasks.ActorStateChangeDiff) (model.Persistable, *visormodel.ProcessingReport, error) {
+	ctx, span := otel.Tracer("").Start(ctx, "ProcessActors")
+	defer span.End()
 	report := &visormodel.ProcessingReport{
-		Height:    int64(pts.Height()),
-		StateRoot: pts.ParentState().String(),
+		Height:    int64(current.Height()),
+		StateRoot: current.ParentState().String(),
 		Status:    visormodel.ProcessingStatusOK,
 	}
 
-	ll := log.With("height", int64(ts.Height()))
-
 	// Filter to just allowed actors
-	actors := make(map[string]tasks.ActorStateChange)
+	actors := make(map[address.Address]tasks.ActorStateChange)
 	for addr, ch := range candidates {
 		if t.extracterMap.Allow(ch.Actor.Code) {
 			actors[addr] = ch
 		}
 	}
+	log.Debugw("processing actor state changes", "height", current.Height(), "")
 
 	data := make(model.PersistableList, 0, len(actors))
 	errorsDetected := make([]*ActorStateError, 0, len(actors))
 	skippedActors := 0
+	ll := log.With("height", int64(current.Height()))
 
 	if len(actors) == 0 {
 		ll.Debugw("no actor state changes found")
@@ -64,40 +68,40 @@ func (t *Task) ProcessActors(ctx context.Context, ts *types.TipSet, pts *types.T
 	start := time.Now()
 	ll.Debugw("found actor state changes", "count", len(actors))
 
-	// Run each task concurrently
 	results := make(chan *ActorStateResult, len(actors))
-	for addr, ch := range actors {
-		go t.runActorStateExtraction(ctx, ts, pts, addr, ch, results)
-	}
+	// closes result when done, runs extraction for each actor in actors concurrently.
+	t.startActorStateExtraction(ctx, current, executed, actors, results)
 
 	// Gather results
-	inFlight := len(actors)
-	for inFlight > 0 {
-		res := <-results
-		inFlight--
-		lla := log.With("height", int64(ts.Height()), "actor", builtin.ActorNameByCode(res.Code), "address", res.Address)
+	for res := range results {
+		select {
+		case <-ctx.Done():
+			return data, report, ctx.Err()
+		default:
+		}
+		lla := log.With("height", int64(current.Height()), "actor", builtin.ActorNameByCode(res.Code), "address", res.Address)
 
 		if res.Error != nil {
-			lla.Errorw("actor returned with error", "error", res.Error.Error())
+			lla.Errorw("actor returned with error", "code", builtin.ActorNameByCode(res.Code), "error", res.Error.Error())
 			errorsDetected = append(errorsDetected, &ActorStateError{
 				Code:    res.Code.String(),
 				Name:    builtin.ActorNameByCode(res.Code),
 				Head:    res.Head.String(),
-				Address: res.Address,
+				Address: res.Address.String(),
 				Error:   res.Error.Error(),
 			})
 			continue
 		}
 
 		if res.SkippedParse {
-			lla.Debugw("skipped actor without extracter")
+			lla.Warn("skipped actor without extractor", "code", builtin.ActorNameByCode(res.Code))
 			skippedActors++
 		}
 
 		data = append(data, res.Data)
 	}
 
-	log.Debugw("completed processing actor state changes", "height", ts.Height(), "success", len(actors)-len(errorsDetected)-skippedActors, "errors", len(errorsDetected), "skipped", skippedActors, "time", time.Since(start))
+	log.Infow("completed processing actor state changes", "height", current.Height(), "success", len(actors)-len(errorsDetected)-skippedActors, "errors", len(errorsDetected), "skipped", skippedActors, "time", time.Since(start))
 
 	if skippedActors > 0 {
 		report.StatusInformation = fmt.Sprintf("did not parse %d actors", skippedActors)
@@ -110,52 +114,66 @@ func (t *Task) ProcessActors(ctx context.Context, ts *types.TipSet, pts *types.T
 	return data, report, nil
 }
 
-func (t *Task) runActorStateExtraction(ctx context.Context, ts *types.TipSet, pts *types.TipSet, addrStr string, ch tasks.ActorStateChange, results chan *ActorStateResult) {
-	ctx, _ = tag.New(ctx, tag.Upsert(metrics.ActorCode, builtin.ActorNameByCode(ch.Actor.Code)))
+func (t *Task) startActorStateExtraction(ctx context.Context, current, executed *types.TipSet, actors tasks.ActorStateChangeDiff, results chan *ActorStateResult) {
+	for addr, ac := range actors {
+		addr := addr
+		ac := ac
 
-	res := &ActorStateResult{
-		Code:    ch.Actor.Code,
-		Head:    ch.Actor.Head,
-		Address: addrStr,
+		t.apwg.Add(1)
+		go func() {
+			defer t.apwg.Done()
+
+			ctx, _ = tag.New(ctx, tag.Upsert(metrics.ActorCode, builtin.ActorNameByCode(ac.Actor.Code)))
+
+			info := ActorInfo{
+				Actor:      ac.Actor,
+				ChangeType: ac.ChangeType,
+				Address:    addr,
+				Current:    current,
+				Executed:   executed,
+			}
+			ae, ok := t.extracterMap.GetExtractors(info.Actor.Code)
+			if !ok {
+				results <- &ActorStateResult{
+					Code:         ac.Actor.Code,
+					Head:         ac.Actor.Head,
+					Address:      addr,
+					SkippedParse: true,
+					Data:         nil,
+				}
+			} else {
+				for _, e := range ae {
+					e := e
+					t.apwg.Add(1)
+					go func() {
+						defer t.apwg.Done()
+
+						res := &ActorStateResult{
+							Code:    ac.Actor.Code,
+							Head:    ac.Actor.Head,
+							Address: addr,
+						}
+						data, err := e.Extract(ctx, info, t.node)
+						if err != nil {
+							res.Error = xerrors.Errorf("failed to extract parsed actor state: %w", err)
+						}
+						res.Data = data
+						results <- res
+					}()
+				}
+			}
+		}()
 	}
-	defer func() {
-		results <- res
+	go func() {
+		t.apwg.Wait()
+		close(results)
 	}()
-
-	addr, err := address.NewFromString(addrStr)
-	if err != nil {
-		res.Error = xerrors.Errorf("failed to parse address: %w", err)
-		return
-	}
-
-	info := ActorInfo{
-		Actor:           ch.Actor,
-		ChangeType:      ch.ChangeType,
-		Address:         addr,
-		ParentStateRoot: ts.ParentState(),
-		Epoch:           ts.Height(),
-		TipSet:          ts,
-		ParentTipSet:    pts,
-	}
-
-	extracter, ok := t.extracterMap.GetExtractor(ch.Actor.Code)
-	if !ok {
-		res.SkippedParse = true
-	} else {
-		// Parse state
-		data, err := extracter.Extract(ctx, info, t.node)
-		if err != nil {
-			res.Error = xerrors.Errorf("failed to extract parsed actor state: %w", err)
-			return
-		}
-		res.Data = data
-	}
 }
 
 type ActorStateResult struct {
 	Code         cid.Cid
 	Head         cid.Cid
-	Address      string
+	Address      address.Address
 	Error        error
 	SkippedParse bool
 	Data         model.Persistable
@@ -172,7 +190,7 @@ type ActorStateError struct {
 // An ActorExtractorMap controls which actor types may be extracted.
 type ActorExtractorMap interface {
 	Allow(code cid.Cid) bool
-	GetExtractor(code cid.Cid) (ActorStateExtractor, bool)
+	GetExtractors(code cid.Cid) ([]ActorStateExtractor, bool)
 }
 
 type ActorExtractorFilter interface {
@@ -180,24 +198,32 @@ type ActorExtractorFilter interface {
 }
 
 // A RawActorExtractorMap extracts all types of actors using basic actor extraction which only parses shallow state.
-type RawActorExtractorMap struct{}
-
-func (RawActorExtractorMap) Allow(code cid.Cid) bool {
-	return true
+type RawActorExtractorMap struct {
+	extractors []ActorStateExtractor
 }
 
-func (RawActorExtractorMap) GetExtractor(code cid.Cid) (ActorStateExtractor, bool) {
-	return ActorExtractor{}, true
+func (r *RawActorExtractorMap) GetExtractors(code cid.Cid) ([]ActorStateExtractor, bool) {
+	return r.extractors, true
+}
+
+func (r *RawActorExtractorMap) Register(ase ActorStateExtractor) {
+	r.extractors = append(r.extractors, ase)
+}
+
+func (r *RawActorExtractorMap) Allow(code cid.Cid) bool {
+	return true
 }
 
 // A TypedActorExtractorMap extracts a single type of actor using full parsing of actor state
 type TypedActorExtractorMap struct {
-	codes *cid.Set
+	codes      *cid.Set
+	extractors []ActorStateExtractor
 }
 
-func NewTypedActorExtractorMap(codes []cid.Cid) *TypedActorExtractorMap {
+func NewTypedActorExtractorMap(codes []cid.Cid, ase ...ActorStateExtractor) *TypedActorExtractorMap {
 	t := &TypedActorExtractorMap{
-		codes: cid.NewSet(),
+		codes:      cid.NewSet(),
+		extractors: ase,
 	}
 	for _, c := range codes {
 		t.codes.Add(c)
@@ -209,9 +235,9 @@ func (t *TypedActorExtractorMap) Allow(code cid.Cid) bool {
 	return t.codes.Has(code)
 }
 
-func (t *TypedActorExtractorMap) GetExtractor(code cid.Cid) (ActorStateExtractor, bool) {
+func (t *TypedActorExtractorMap) GetExtractors(code cid.Cid) ([]ActorStateExtractor, bool) {
 	if !t.Allow(code) {
 		return nil, false
 	}
-	return GetActorStateExtractor(code)
+	return t.extractors, true
 }
